+++ release-files.nix
@@ -1,0 +1,39 @@
+# This file is used by .github/workflows/release.yaml
+# to create the files to be uploaded
+
+# It imports default.nix both for linux and darwin, thus it cannot be part of
+# it.
+
+{ releaseVersion ? "latest" }:
+let
+  nixpkgs = import ./nix { };
+  linux = import ./default.nix { system = "x86_64-linux"; internal = true; inherit releaseVersion; };
+  darwin = import ./default.nix { system = "x86_64-darwin"; internal = true; inherit releaseVersion; };
+
+  as_tarball = dir: derivations:
+    nixpkgs.runCommandNoCC "motoko-${releaseVersion}.tar.gz" {
+      allowedRequisites = [];
+    } ''
+      tmp=$(mktemp -d)
+      ${nixpkgs.lib.concatMapStringsSep "\n" (d: "cp -v ${d}/bin/* $tmp") derivations}
+      chmod 0755 $tmp/*
+      tar -czf "$out" -C $tmp/ .
+    '';
+
+  as_js = name: derivation:
+    nixpkgs.runCommandNoCC "${name}-${releaseVersion}.js" {
+      allowedRequisites = [];
+    } ''
+      cp -v ${derivation}/bin/* $out
+    '';
+
+  release =
+    nixpkgs.runCommandNoCC "motoko-release-${releaseVersion}" {} ''
+      mkdir $out
+      cp ${as_tarball "x86_64-linux" (with linux; [ mo-ide mo-doc moc ])} $out/motoko-linux64-${releaseVersion}.tar.gz
+      cp ${as_tarball "x86_64-darwin" (with darwin; [ mo-ide mo-doc moc ])} $out/motoko-macos-${releaseVersion}.tar.gz
+      cp ${as_js "moc" linux.js.moc} $out/moc-${releaseVersion}.js
+      cp ${as_js "moc-interpreter" linux.js.moc_interpreter} $out/moc-interpreter-${releaseVersion}.js
+    '';
+in
+release